Feature: chord keyboard (#2385)
* Add initial chord keyboard

Basic proof of concept chord keyboard
just a few chords and horizontal scrolling works
it has colored columns, but those don't change with horizontal scrolling
yet

* Now displays chord name

The chord name is now displayed when pressing the chord
The chord keyboard class is entirely responsible for displaying
There is an override in `instrument_clip_minder` to not display note
name when the keyboard type is chord keyboard and keyboard is active
Also, had to reimplement some of `noteCodeToString` because on some
of the columns I was getting garbage characters (F#@ for example instead
of F#) or if you played two chords together, the it wouldn't display
properly.

* Fix crashing bug from vtable looking on constructor and destructor

* Move chord structs to chords.h file

* Add vertical encoder to move up to more chords

* Add voicings with x Enc

Keyboards now know if the x Enc is being pressed when the
handle horizontal enc funtion is being called and what notes
are being pressed. This allowed me to change voicings on
the chords being pressed.
The keyboards currently implemented had their signature changed
but nothing about their implementation changed

Also added Chords class instead of Chords struct
This allowed me to better do bounds checks for voicings.
So if some chord doesn't have as many voicings,
it will just return the last defined voicing.

* Formatting

* Add chord definitions as seperate consts

Also add interval consts for easier chord defining

* Clean up comments

* Simplify logic on some of encoder checks

* Add check for NON interval so chords can more easily be rootless

* Can now add sub names to Voicings

* Add community_features section

* Update Changelog

* Add ChordList get voicing bounds check test

* Move more checks into ChordList class, add tests

* Chord and Velocity Keyboards display CANT when entering scale

* Cleanups to drawChordName

* Add color rows on chord keyboard to help with navigation

* Add dim, aug, and m7b5 chords

* Move Chord Keyboard behind community runtime menu flag

+++ src/deluge/gui/ui/keyboard/keyboard_screen.cpp
@@ -14,7 +14,6 @@
  * You should have received a copy of the GNU General Public License along with this program.
  * If not, see <https://www.gnu.org/licenses/>.
  */
-
 #include "gui/ui/keyboard/keyboard_screen.h"
 #include "definitions_cxx.hpp"
 #include "extern.h"
@@ -43,6 +42,7 @@
 #include <cstring>
 
 #include "gui/ui/keyboard/layout.h"
+#include "gui/ui/keyboard/layout/chord_keyboard.h"
 #include "gui/ui/keyboard/layout/in_key.h"
 #include "gui/ui/keyboard/layout/isomorphic.h"
 #include "gui/ui/keyboard/layout/norns.h"
@@ -55,6 +55,7 @@ namespace deluge::gui::ui::keyboard {
 layout::KeyboardLayoutIsomorphic keyboardLayoutIsomorphic{};
 layout::KeyboardLayoutVelocityDrums keyboardLayoutVelocityDrums{};
 layout::KeyboardLayoutInKey keyboardLayoutInKey{};
+layout::KeyboardLayoutChord keyboardLayoutChord{};
 layout::KeyboardLayoutNorns keyboardLayoutNorns{};
 KeyboardLayout* layoutList[KeyboardLayoutType::KeyboardLayoutTypeMaxElement + 1] = {0};
 
@@ -62,6 +63,7 @@ KeyboardScreen::KeyboardScreen() {
 	layoutList[KeyboardLayoutType::KeyboardLayoutTypeIsomorphic] = (KeyboardLayout*)&keyboardLayoutIsomorphic;
 	layoutList[KeyboardLayoutType::KeyboardLayoutTypeDrums] = (KeyboardLayout*)&keyboardLayoutVelocityDrums;
 	layoutList[KeyboardLayoutType::KeyboardLayoutTypeInKey] = (KeyboardLayout*)&keyboardLayoutInKey;
+	layoutList[KeyboardLayoutType::KeyboardLayoutTypeChords] = (KeyboardLayout*)&keyboardLayoutChord;
 	layoutList[KeyboardLayoutType::KeyboardLayoutTypeNorns] = (KeyboardLayout*)&keyboardLayoutNorns;
 
 	memset(&pressedPads, 0, sizeof(pressedPads));
@@ -405,8 +407,12 @@ ActionResult KeyboardScreen::buttonAction(deluge::hid::Button b, bool on, bool i
 
 	// Scale mode button
 	if (b == SCALE_MODE) {
-		if (getCurrentOutputType() == OutputType::KIT) {
-			return ActionResult::DEALT_WITH; // Kits can't do scales!
+		if ((getCurrentOutputType() == OutputType::KIT)
+		    || (getCurrentInstrumentClip()->keyboardState.currentLayout
+		        == KeyboardLayoutType::KeyboardLayoutTypeChords)) {
+			// Kits and Chords can't do scales!
+			display->displayPopup(deluge::l10n::get(deluge::l10n::String::STRING_FOR_KEYBOARD_VIEW_CANT_ENTER_SCALE));
+			return ActionResult::DEALT_WITH;
 		}
 
 		actionLogger.deleteAllLogs(); // Can't undo past this!
@@ -560,6 +566,11 @@ ActionResult KeyboardScreen::buttonAction(deluge::hid::Button b, bool on, bool i
 		requestRendering();
 	}
 
+	// store if the user is holding the x encoder
+	else if (b == X_ENC) {
+		xEncoderActive = on;
+	}
+
 	else {
 		requestRendering();
 		ActionResult result = InstrumentClipMinder::buttonAction(b, on, inCardRoutine);
@@ -598,7 +609,8 @@ ActionResult KeyboardScreen::verticalEncoderAction(int32_t offset, bool inCardRo
 ActionResult KeyboardScreen::horizontalEncoderAction(int32_t offset) {
 
 	layoutList[getCurrentInstrumentClip()->keyboardState.currentLayout]->handleHorizontalEncoder(
-	    offset, (Buttons::isShiftButtonPressed() && isUIModeWithinRange(padActionUIModes)));
+	    offset, (Buttons::isShiftButtonPressed() && isUIModeWithinRange(padActionUIModes)), pressedPads,
+	    xEncoderActive);
 
 	if (isUIModeWithinRange(padActionUIModes)) {
 		evaluateActiveNotes();
@@ -623,8 +635,11 @@ void KeyboardScreen::selectLayout(int8_t offset) {
 			nextLayout = 0;
 		}
 
-		if (runtimeFeatureSettings.get(RuntimeFeatureSettingType::DisplayNornsLayout) == RuntimeFeatureStateToggle::Off
-		    && nextLayout == KeyboardLayoutType::KeyboardLayoutTypeNorns) {
+		if ((runtimeFeatureSettings.get(RuntimeFeatureSettingType::DisplayNornsLayout) == RuntimeFeatureStateToggle::Off
+		     && nextLayout == KeyboardLayoutType::KeyboardLayoutTypeNorns)
+		    || (runtimeFeatureSettings.get(RuntimeFeatureSettingType::DisplayChordKeyboard)
+		            == RuntimeFeatureStateToggle::Off
+		        && nextLayout == KeyboardLayoutType::KeyboardLayoutTypeChords)) {
 			// Don't check the next conditions, this one is already lost
 		}
 		else if (getCurrentOutputType() == OutputType::KIT && layoutList[nextLayout]->supportsKit()) {
@@ -665,7 +680,8 @@ void KeyboardScreen::selectLayout(int8_t offset) {
 	}
 
 	// Ensure scroll values are calculated in bounds
-	layoutList[getCurrentInstrumentClip()->keyboardState.currentLayout]->handleHorizontalEncoder(0, false);
+	layoutList[getCurrentInstrumentClip()->keyboardState.currentLayout]->handleHorizontalEncoder(0, false, pressedPads,
+	                                                                                             xEncoderActive);
 
 	// Precalculate because changing instruments can change pad colours
 	layoutList[getCurrentInstrumentClip()->keyboardState.currentLayout]->precalculate();
@@ -691,14 +707,16 @@ void KeyboardScreen::selectEncoderAction(int8_t offset) {
 			}
 		}
 		display->displayPopup(noteName, 3, false, (noteCodeIsSharp[newRootNote] ? 0 : 255));
-		layoutList[getCurrentInstrumentClip()->keyboardState.currentLayout]->handleHorizontalEncoder(0, false);
+		layoutList[getCurrentInstrumentClip()->keyboardState.currentLayout]->handleHorizontalEncoder(
+		    0, false, pressedPads, xEncoderActive);
 		layoutList[getCurrentInstrumentClip()->keyboardState.currentLayout]->precalculate();
 		requestRendering();
 	}
 	else {
 		InstrumentClipMinder::selectEncoderAction(offset);
 		// Ensure scroll values are calculated in bounds
-		layoutList[getCurrentInstrumentClip()->keyboardState.currentLayout]->handleHorizontalEncoder(0, false);
+		layoutList[getCurrentInstrumentClip()->keyboardState.currentLayout]->handleHorizontalEncoder(
+		    0, false, pressedPads, xEncoderActive);
 		layoutList[getCurrentInstrumentClip()->keyboardState.currentLayout]->precalculate();
 		requestRendering();
 	}
@@ -737,7 +755,8 @@ void KeyboardScreen::openedInBackground() {
 	getCurrentInstrumentClip()->onKeyboardScreen = true;
 
 	// Ensure scroll values are calculated in bounds
-	layoutList[getCurrentInstrumentClip()->keyboardState.currentLayout]->handleHorizontalEncoder(0, false);
+	layoutList[getCurrentInstrumentClip()->keyboardState.currentLayout]->handleHorizontalEncoder(0, false, pressedPads,
+	                                                                                             xEncoderActive);
 	layoutList[getCurrentInstrumentClip()->keyboardState.currentLayout]->precalculate();
 	requestRendering(); // This one originally also included sidebar, the other ones didn't
 }
